pytest: Migrate the gossip related tests to the new fixture model
